requirements.txt
#torch==2.0.1+cu118
torch-ema==0.3
torch-fidelity==0.3.0
torch-summary==1.4.5
#torchaudio==2.0.2+cu118
torchcde==0.2.5
torchcfm==1.0.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchmetrics==1.2.1
torchsde==0.2.6
#torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trampoline==0.1.2
transformers==4.43.4
triton==2.0.0
typeguard==2.13.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.5
uri-template==1.3.0
urllib3==2.1.0
visdom==0.2.4
wandb==0.16.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
absl-py==2.0.0
accelerate==0.29.2
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.4.1
anndata==0.10.3
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.4
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
#basicsr==1.4.2
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.2
calflops==0.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
cmake==3.25.0
colored-traceback==0.3.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dominate==2.9.1
easydict==1.13
entrypoints==0.4
#eot-benchmark==1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==4.7.1
geotorch==0.3.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.1
google-auth-oauthlib==1.1.0
GPUtil==1.4.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
images2gif==1.0.1
importlib_metadata==7.1.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.1.5
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-bolts==0.7.0
lightning-utilities==0.10.0
lit==15.0.7
llvmlite==0.41.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multipledispatch==1.0.0
natsort==8.4.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
nip-config==0.10.3
notebook==7.1.2
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
papermill==2.5.0
parso==0.8.3
pastel==0.2.1
patsy==0.5.4
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.1.0
poethepoet==0.10.0
portalocker==2.10.1
POT==0.9.1
prefetch-generator==1.0.3
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pynndescent==0.5.11
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-fid==0.3.0
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.7.24
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.3
scanpy==1.9.6
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
scprep==1.2.3
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.38.0
session-info==1.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
stdlib-list==0.10.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.3
